+++ packages/jaspr/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added support for rendering `svg`.
 - Added new `Colors.transparent` color.
 - Added `Unit.auto`, `Unit.vw` and `Unit.vh` for responsive styling.
+- Added `StyleRule.fontFace` to add external font files.
 
 ## 0.9.3